camera/requirements.txt
numpy==1.19.5
RPi.GPIO==0.7.0
websockets==10.2